+++ archon-ui-main/package.json
@@ -7,6 +7,7 @@
     "dev": "npx vite",
     "build": "npx vite build",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
+    "lint:files": "eslint --ext .js,.jsx,.ts,.tsx",
     "preview": "npx vite preview",
     "test": "vitest",
     "test:ui": "vitest --ui",
