+++ archon-ui-main/src/features/projects/components/tasks/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Task Management Components
+ *
+ * Simplified and refactored task components following vertical slice architecture.
+ * Removed complex flip animations and over-engineering for better maintainability.
+ */
+
+export { TaskCard } from "./TaskCard";
+export { TaskCardActions } from "./TaskCardActions";
+export type { TaskCardProps } from "./TaskCard";
